SeachAlbumsSpecialRequest: Move To "special" Package
