+++ src/test/fixtures/requests/data/search/simplified/special/SearchAlbumsSpecialRequest.json
